Updated all the dependencies to the latest.
Tried bootstrap4, but it was too much effort to go through the whole
migration + react-bootstrap only supports bootstrap 3.

For this reason, i am staying with less. Tried sass, but for comfort
reason (i dont want to do bootstrap-sass and some weird bindings etc.)

There are some performance issues to fix :/

Updates the gulpfile to gulpfile.babel.js to use es6 syntax in the
gulpfile.

Package.json needs to be cleaned.