Fixed a bug regarding modals being ads
When you click on the info button for a course it will open the
modal. But the the close state was only being saved on the modal itself
not at the parent. So when the modal gets closed, the parent will still
think that the modal should be open. So when the parent gets rerendered
it will pop up the modal again.