STarted with the view of generating your ISP
For now commented out the year view and replaced it with the ISP select
view. For now it generates the fields based on a constant file with
given options like: name, maxEC, maxCourses etc.

You can drag n drop the courses from your selected course to the isp
fields.

Still a lot to do:
 - Introduce states: to switch from year view to select view
 - Check for constraints and visiualize it for the user
 - save/load
 - generating an html page which is printable