These changes reflect the change in the allcourses.json See prev commit what happened to allcourses.json Now only the course tree object will be passed through all the components. if a component wants more information of the course object it should call CourseCtrl.get
Also changed a performance issue (hopefully this will make the initial load faster) it will now not render the coursetree components if they are not visible.
the tradeoff is that when toggling the visibility it will now render the course tree instead of setting the visibility.
i will test if this will increase the performance or not